+++ src/app/_provider/JotaiProvider.tsx
@@ -1,8 +1,8 @@
-"use client"
+'use client'
 
-import { Provider } from "jotai"
-import React from "react"
-import { HasChildren } from "../util/interface"
+import { Provider } from 'jotai'
+import React from 'react'
+import { HasChildren } from '../util/interface'
 
 export default function JotaiProvider({ children }: HasChildren) {
   return <Provider>{children}</Provider>
