+++ src/app/components/layout/BasicContainer.tsx
@@ -1,5 +1,5 @@
-import { HasChildren } from "@/app/util/interface"
-import { FC } from "react"
+import { HasChildren } from '@/app/util/interface'
+import { FC } from 'react'
 
 const BasicContainder: FC<HasChildren> = ({ children }) => {
   return <div className="w-[1700px] mr-auto ml-auto">{children}</div>
