+++ src/app/layout.tsx
@@ -1,11 +1,11 @@
-import type { Metadata } from "next"
-import "./globals.css"
-import JotaiProvider from "./_provider/JotaiProvider"
-import TopNavigation from "./components/templates/TopNavigation"
+import type { Metadata } from 'next'
+import './globals.css'
+import JotaiProvider from './_provider/JotaiProvider'
+import TopNavigation from './components/templates/TopNavigation'
 
 export const metadata: Metadata = {
-  title: "Nextjs15",
-  description: "Study Nextj15 App",
+  title: 'Nextjs15',
+  description: 'Study Nextj15 App',
 }
 
 export default function RootLayout({
