+++ src/app/util/interface.ts
@@ -1,4 +1,4 @@
-import React from "react"
+import React from 'react'
 
 export interface HasChildren {
   children: React.ReactNode
